+++ brue/_utils.py
@@ -2,7 +2,7 @@
 import os, sys, subprocess
 from glob import glob
 
-def transpile(source:str, dest:str, remove_settings:bool = False, remove_dest:bool = False, search_css:bool = False, verbose:bool = True) -> str:
+def transpile(source:str, dest:str, remove_dest:bool = False, search_css:bool = False, verbose:bool = True) -> str:
     subprocess.call([ sys.executable, "-m", "metapensiero.pj", source, "-o", dest ], stdout = sys.stdout if verbose else subprocess.DEVNULL)
     with open(dest, "r", encoding = "utf-8-sig") as rt:
         lines = []
@@ -17,8 +17,7 @@ def transpile(source:str, dest:str, remove_settings:bool = False, remove_dest:bo
 
         rtt = "\n".join(lines).strip()
 
-    if remove_settings:
-        rtt = rtt.split("_pj_snippets(_pj);")[-1]
+    rtt = rtt.split("_pj_snippets(_pj);")[-1]
 
     if search_css:
         try:
@@ -41,7 +40,7 @@ def transpile(source:str, dest:str, remove_settings:bool = False, remove_dest:bo
     return rtt
 
 
-def transpile_directory(source_patterns:str, dest_dir:str, ext:str = "js", merge_file:str = None, remove_settings:bool = False, verbose:bool = False):
+def transpile_directory(source_patterns:str, dest_dir:str, ext:str = "js", merge_file:str = None, verbose:bool = False):
     texts = []
 
     dir_pattern = os.path.dirname(source_patterns)
@@ -50,7 +49,7 @@ def transpile_directory(source_patterns:str, dest_dir:str, ext:str = "js", merge
     files_to_transpile = glob(source_patterns) + glob(os.path.join(dir_pattern, "**", file_pattern))
     for idx, file in enumerate(files_to_transpile):
         file_name = os.path.splitext(os.path.basename(file))[0]
-        texts.append(transpile(file, os.path.join(dest_dir, f"{file_name}.{ext}"), remove_settings = (not idx == 0 or remove_settings), remove_dest = merge_file is None, search_css = True, verbose = verbose))
+        texts.append(transpile(file, os.path.join(dest_dir, f"{file_name}.{ext}"), remove_dest = merge_file is None, search_css = True, verbose = verbose))
 
     if merge_file is not None:
         with open(os.path.join(dest_dir, "script.js"), "a", encoding = "utf-8-sig") as sw: